requirements.txt
langchain
python-dotenv
langchain-core
langchain-community
langchain_huggingface
langchain-groq
langgraph
langgraph-cli[inmem]
youtube_transcript_api
IPython
typing-extensions